+++ client/gutenberg/extensions/presets/jetpack/editor.js
@@ -14,6 +14,7 @@ import * as MapBlock from 'gutenberg/extensions/map';
 import * as PublicizeBlock from 'gutenberg/extensions/publicize';
 import * as RelatedPostsBlock from 'gutenberg/extensions/related-posts';
 import * as SimplePaymentsBlock from 'gutenberg/extensions/simple-payments';
+import * as SubscriptionsBlock from 'gutenberg/extensions/subscriptions';
 import * as TiledGalleryBlock from 'gutenberg/extensions/tiled-gallery';
 import * as VRBlock from 'gutenberg/extensions/vr';
 import { isEnabled } from 'config';
@@ -25,6 +26,7 @@ export default [
 	MapBlock,
 	PublicizeBlock,
 	SimplePaymentsBlock,
+	SubscriptionsBlock,
 	...( isEnabled( 'jetpack/blocks/beta' )
 		? [ RelatedPostsBlock, TiledGalleryBlock, VRBlock ]
 		: [] ),
